requirements.txt
#torch
#torchvision
#mmcv-full
wandb
prettytable
mmcv-full==1.3.9
mmdet==2.16.0